# openapi/aisi.v1.yaml
openapi: 3.0.3
info:
  title: AISI / Dakwah Depok API
  version: 1.0.0
  description: |
    Kontrak SSOT API AISI (Go).
    Envelope sukses: `{ success: true, data, message?, pagination? }`
    Envelope error: `{ success: false, message }`
servers:
  - url: /api/v1
paths:
...
  /users/leaderboard:
    get:
      operationId: usersLeaderboard
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /users/me:
    get:
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
    put:
      operationId: updateMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /users:
    get:
      operationId: listUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /users/{id}:
    get:
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
    put:
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
    delete:
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
...